Fix all compiler warnings and add -Wpedantic to build flags (#2)
* Initial plan

* Fix all compiler warnings and add -Wpedantic flag to build

Agent-Logs-Url: https://github.com/acoto87/shl/sessions/74401d63-407c-40c4-bbeb-c119acfaea52

* Switch container count and capacity fields to int32_t

Agent-Logs-Url: https://github.com/acoto87/shl/sessions/a2cd5e06-2bf0-4276-89ff-fd0341b3e244

* Retrieve value before removing item from list

---------

+++ tests/binary_heap_test.c
@@ -194,7 +194,7 @@ void referenceTypeTest()
     printf("--- Start test 2: peek min object ---\n");
     start = getTime();
     const char* peek = SHeapPeek(&heap);
-    assert(strlen(peek) == min);
+        assert(strlen(peek) == (size_t)min);
     end = getTime();
     printf("Time: %.2f seconds\n", end - start);
     printf("--- End test 2:  peek min object ---\n");
